+++ src/Config.h
@@ -29,6 +29,8 @@ struct Preset {
     bool use_manual_slip;
     int bottoming_method;
     float scrub_drag_gain;
+    // New Params (v0.4.11)
+    float rear_align_effect;
     
     // Apply this preset to an engine instance
     void Apply(FFBEngine& engine) const {
@@ -52,6 +54,7 @@ struct Preset {
         engine.m_use_manual_slip = use_manual_slip;
         engine.m_bottoming_method = bottoming_method;
         engine.m_scrub_drag_gain = scrub_drag_gain;
+        engine.m_rear_align_effect = rear_align_effect;
     }
 };